Fix a 64-bit pointer truncated through Lua; the crash it caused is gone
The fault address named the bug. 0xfffffffff6071b60 is 0xf6071b60 sign-extended,
which happens exactly one way: a pointer carried in a 32-bit signed int. The
whole round trip truncates at both ends.

  WorldClient, sending a selected unit's handle to a script:
      NStr::Format( "%d,", reinterpret_cast<int>( ... ) )
  Scripts::ReturnScriptIDs, reading it back:
      const int nPtr = script.GetObject( i );
      IRefCount *pObj = reinterpret_cast<IRefCount*>( nPtr );
      dynamic_cast<IUpdatableObj*>( pObj )        <- reads a vtable at that
                                                     address; process dies

On 32-bit Windows an address and an int are the same width and this was
harmless. Here the top half is dropped on the way out and the remainder is
sign-extended on the way back.

Fixed by carrying the value at full width on both sides. Lua stores numbers as
doubles -- 53 bits of mantissa, which holds every address this platform hands
out -- so the transport did not have to change. Script.h gained GetExactNumber
because everything it already offered narrowed: GetInteger to 32 bits, GetNumber
to a float's 24-bit mantissa.

Correct on the 32-bit build too, so it is not guarded behind the port. This is
the fifth truncation of its kind found here, after the four in CTextureLock and
the GFX helpers.

Verified by repeating the exact sequence that killed the process -- select a
tank, order it at a building. It now runs through: process alive, 57.7 fps, 126
draws, no signal in the log.

The collision measurement that was being attempted when this surfaced is still
not taken. The obstacle test needs the tank in frame at the end and it was not;
that is unfinished, not passed.

+++ Sources/src/GameTT/WorldClient.cpp
@@ -750,8 +750,17 @@ void CWorldClient::Update( const NTimer::STime &currTime )
 			{
 				std::string szCmd = "ReturnScriptIDs( ";
 
+				// The handle a script gets for a unit is the object's address, and
+				// %d with a cast to int throws away the top half of it on a
+				// 64-bit build. ReturnScriptIDs turns the number back into a
+				// pointer, so what came back was the low half sign-extended --
+				// 0xfffffffff6071b60 -- and the dynamic_cast on it killed the
+				// process. Printed in full here and read in full on the other
+				// side; Lua carries it in a double, which holds every address
+				// this platform hands out.
 				for ( CMapObjectsList::const_iterator it = selList.begin(); it != selList.end(); ++it )
-					szCmd += NStr::Format( "%d,", reinterpret_cast<int>((*it)->pAIObj.GetPtr()) );
+					szCmd += NStr::Format( "%llu,",
+					                       (unsigned long long)reinterpret_cast<uintptr_t>( (*it)->pAIObj.GetPtr() ) );
 				szCmd.resize( szCmd.size() - 1 );
 				szCmd += " )";
